✨ (UI): Fetching directory structure on load
Now using IPC and useEffect we are fetching the directory structure from the main thread of electron and displaying it using React on the client-side. This all happens automatically and doesn't hang the application. This might need to use suspense & fallback in future when the directory structure gets more complex